Revert "python: Run `check_bazel_python` at build-time"
