+++ tools/workspace/python/BUILD.bazel
@@ -4,7 +4,11 @@
 # neighboring *.bzl file can be loaded elsewhere.
 
 load("//tools/lint:lint.bzl", "add_lint_tests")
+load("@drake//tools/skylark:drake_py.bzl", "drake_py_unittest")
 
-add_lint_tests(
-    python_lint_extra_srcs = ["check_bazel_python.py"],
+drake_py_unittest(
+    name = "python_bin_test",
+    deps = ["@python//:bazel_python_actionenv"],
 )
+
+add_lint_tests()
