+++ tools/workspace/python/repository.bzl
@@ -58,14 +58,16 @@ def _repository_python_info(repository_ctx):
         os_key = os_result.distribution + ":" + os_result.ubuntu_release
     versions_supported = _VERSION_SUPPORT_MATRIX[os_key]
 
-    # Bazel does not easily expose its --python_path to repository rules
-    # (analysis phase). We must use a workaround as Tensorflow does in
-    # `python_configure.bzl` (https://git.io/fx4Pp). We check for consistency
-    # during the build (execution) phase using `bazel_python_is_valid`.
+    # Bazel does not easily expose its --python_path to repository rules (e.g.
+    # the environment is unaffected). We must use a workaround as Tensorflow
+    # does in `python_configure.bzl` (https://git.io/fx4Pp).
+    # N.B. Unfortunately, it does not seem possible to get Bazel's Python
+    # interpreter during a repository rule, thus we can only catch mismatch
+    # issues via `//tools/workspace/python:py/python_bin_test`.
+    # TODO(eric.cousineau): Make this an error once `.bazelrc` stops using
+    # `try-import` for configuration.
     python_path = repository_ctx.os.environ.get("DRAKE_PYTHON_BIN_PATH")
     if python_path == None:
-        # TODO(eric.cousineau): Make this an error once `.bazelrc` stops using
-        # `try-import` for configuration.
         if os_result.is_macos:
             python_path = "/usr/local/bin/python{}".format(
                 versions_supported[0],
@@ -195,12 +197,8 @@ PYTHON_SITE_PACKAGES_RELPATH = "{site_packages_relpath}"
         content = skylark_content,
         executable = False,
     )
-    repository_ctx.symlink(
-        Label("@drake//tools/workspace/python:check_bazel_python.py"),
-        "_check_bazel_python.py",
-    )
     repository_ctx.file(
-        "_bazel_python_actionenv.py",
+        "bazel_python_actionenv.py",
         content = skylark_content,
         executable = False,
     )
@@ -222,9 +220,6 @@ headers = glob(
 
 cc_library(
     name = "python_headers",
-    # Depend on a Python configuration sanity check for anything that wishes to
-    # generate bindings. See `genrule` below for more information.
-    data = [":bazel_python_is_valid"],
     hdrs = headers,
     includes = {},
     visibility = ["//visibility:private"],
@@ -244,28 +239,14 @@ cc_library(
     visibility = ["//visibility:public"],
 )
 
-# See `genrule` below.
-py_binary(
-    name = "check_bazel_python",
-    main = "_check_bazel_python.py",
-    srcs = [
-        "_check_bazel_python.py",
-        "_bazel_python_actionenv.py",
-    ],
+py_library(
+    name = "bazel_python_actionenv",
+    srcs = ["bazel_python_actionenv.py"],
     imports = ["."],
-    visibility = ["//visibility:private"],
+    visibility = ["//visibility:public"],
+    testonly = 1,
 )
-
-# Place this test as a `genrule` to (a) test at build time and (b) be able to
-# access Bazel's Python interpreter from a `py_binary` used in `tools`.
-genrule(
-    name = "bazel_python_is_valid",
-    outs = [".bazel_python_is_valid"],
-    cmd = "$(location :check_bazel_python) > $@",
-    tools = [":check_bazel_python"],
-    visibility = ["//visibility:private"],
-)
-""".format(includes, linkopts, linkopts_direct_link)
+    """.format(includes, linkopts, linkopts_direct_link)
 
     repository_ctx.file(
         "BUILD.bazel",
